Fix: Move __getattr__ shim for Anchor and Package to importlib_resources.__init__, since that's where it's eagerly imported from.
